requirements.txt
langgraph
langchain
langchain-core
langchain-community
langchain-openai
langchain-qdrant
qdrant-client
sentence-transformers
langchain-huggingface
pydantic
rich
python-dotenv
typing-extensions
